# api/swagger/swagger.yaml
basePath: /api/1.0/
definitions:
...
  github_com_openinfradev_tks-api_pkg_domain.FilterResponse:
    properties:
      column:
        type: string
      operator:
        type: string
      releation:
        type: string
      values:
        items:
          type: string
        type: array
    type: object
  github_com_openinfradev_tks-api_pkg_domain.FindIdRequest:
    properties:
      code:
        type: string
      email:
        type: string
      organizationId:
        type: string
      userName:
        type: string
    required:
    - code
    - email
    - organizationId
    - userName
    type: object
  github_com_openinfradev_tks-api_pkg_domain.FindIdResponse:
    properties:
      accountId:
        type: string
    type: object
  github_com_openinfradev_tks-api_pkg_domain.FindPasswordRequest:
    properties:
      accountId:
        type: string
      code:
        type: string
      email:
        type: string
      organizationId:
        type: string
      userName:
        type: string
    required:
    - accountId
    - code
    - email
    - organizationId
    - userName
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.GetPolicyNotificationResponse:
    properties:
      policyNotification:
        $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.PolicyNotificationResponse'
    type: object
  github_com_openinfradev_tks-api_pkg_domain.GetPolicyNotificationsResponse:
    properties:
      pagination:
        $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.PaginationResponse'
      policyNotifications:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.PolicyNotificationResponse'
        type: array
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.LoginRequest:
    properties:
      accountId:
        type: string
      organizationId:
        type: string
      password:
        type: string
    required:
    - accountId
    - organizationId
    - password
    type: object
  github_com_openinfradev_tks-api_pkg_domain.LoginResponse:
    properties:
      user:
        properties:
          accountId:
            type: string
          department:
            type: string
          name:
            type: string
          organization:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.OrganizationResponse'
          passwordExpired:
            type: boolean
          roles:
            items:
              $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SimpleRoleResponse'
            type: array
          token:
            type: string
        type: object
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.OrganizationResponse:
    properties:
      admin:
        $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SimpleUserResponse'
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      policyTemplates:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SimplePolicyTemplateResponse'
        type: array
      primaryClusterId:
        type: string
      stackCount:
        type: integer
      stackTemplates:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SimpleStackTemplateResponse'
        type: array
      status:
        type: string
      statusDesc:
        type: string
      systemNotificationTemplates:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SimpleSystemNotificationTemplateResponse'
        type: array
      updatedAt:
        type: string
    type: object
  github_com_openinfradev_tks-api_pkg_domain.PaginationResponse:
    properties:
      filters:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.FilterResponse'
        type: array
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sortColumn:
        type: string
      sortOrder:
        type: string
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.PolicyNotificationResponse:
    properties:
      cluster:
        $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SimpleClusterResponse'
      createdAt:
        type: string
      grafanaUrl:
        type: string
      id:
        type: string
      messageActionProposal:
        type: string
      messageContent:
        type: string
      messageTitle:
        type: string
      notificationType:
        type: string
      organizationId:
        type: string
      policyName:
        type: string
      rawData:
        type: string
      read:
        type: boolean
      severity:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.SimpleClusterResponse:
    properties:
      id:
        type: string
      name:
        type: string
      organizationId:
        type: string
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.SimplePolicyTemplateResponse:
    properties:
      description:
        example: 이 정책은 ...
        type: string
      id:
        example: d98ef5f1-4a68-4047-a446-2207787ce3ff
        type: string
      templateName:
        example: 필수 Label 검사
        type: string
      type:
        enum:
        - tks
        - organization
        example: tks
        type: string
      version:
        example: v1.0.1
        type: string
    type: object
  github_com_openinfradev_tks-api_pkg_domain.SimpleRoleResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  github_com_openinfradev_tks-api_pkg_domain.SimpleStackTemplateResponse:
    properties:
      cloudService:
        type: string
      description:
        type: string
      id:
        type: string
      kubeType:
        type: string
      kubeVersion:
        type: string
      name:
        type: string
      services:
        items:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.SimpleStackTemplateServiceResponse'
        type: array
      template:
        type: string
    type: object
  github_com_openinfradev_tks-api_pkg_domain.SimpleStackTemplateServiceResponse:
    properties:
      type:
        type: string
    type: object
  github_com_openinfradev_tks-api_pkg_domain.SimpleSystemNotificationTemplateResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_openinfradev_tks-api_pkg_domain.SimpleUserResponse:
    properties:
      accountId:
        type: string
      department:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  github_com_openinfradev_tks-api_pkg_domain.VerifyIdentityForLostIdRequest:
    properties:
      email:
        type: string
      organizationId:
        type: string
      userName:
        type: string
    required:
    - email
    - organizationId
    - userName
    type: object
  github_com_openinfradev_tks-api_pkg_domain.VerifyIdentityForLostIdResponse:
    properties:
      validityPeriod:
        type: string
    type: object
  github_com_openinfradev_tks-api_pkg_domain.VerifyIdentityForLostPasswordRequest:
    properties:
      accountId:
        type: string
      email:
        type: string
      organizationId:
        type: string
      userName:
        type: string
    required:
    - accountId
    - email
    - organizationId
    - userName
    type: object
  github_com_openinfradev_tks-api_pkg_domain.VerifyIdentityForLostPasswordResponse:
    properties:
      validityPeriod:
        type: string
    type: object
...
  github_com_openinfradev_tks-api_pkg_httpErrors.RestError:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
      text:
        type: string
    type: object
host: tks-api-dev.taco-cat.xyz
info:
  contact:
    email: taekyu.kang@sk.com
    name: taekyu.kang@sk.com
  description: This is backend api service for tks platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: tks-api service
  version: "1.0"
paths:
...
  /auth/find-id/code:
    post:
      consumes:
      - application/json
      description: This API allows users to verify their identity for lost id by submitting
        required information
      parameters:
      - description: Request body for verifying identity for lost id including {organization
          ID, email, username}
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.VerifyIdentityForLostIdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.VerifyIdentityForLostIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_httpErrors.RestError'
      summary: Request to verify identity for lost id
      tags:
      - Auth
  /auth/find-id/verification:
    post:
      consumes:
      - application/json
      description: This API allows users to find their account ID by submitting required
        information
      parameters:
      - description: Request body for finding the account ID including {organization
          ID, email, username, 6 digit code}
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.FindIdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.FindIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_httpErrors.RestError'
      summary: Request to find forgotten ID
      tags:
      - Auth
  /auth/find-password/code:
    post:
      consumes:
      - application/json
      description: This API allows users to verify their identity for lost password
        by submitting required information
      parameters:
      - description: Request body for verifying identity for lost password including
          {organization ID, email, username, Account ID}
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.VerifyIdentityForLostPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.VerifyIdentityForLostPasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_httpErrors.RestError'
      summary: Request to verify identity for lost password
      tags:
      - Auth
  /auth/find-password/verification:
    post:
      consumes:
      - application/json
      description: This API allows users to reset their forgotten password by submitting
        required information
      parameters:
      - description: Request body for finding the password including {organization
          ID, email, username, Account ID, 6 digit code}
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.FindPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_httpErrors.RestError'
      summary: Request to find forgotten password
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: login
      parameters:
      - description: account info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: user detail
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.LoginResponse'
      summary: login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: logout
      produces:
      - application/json
      responses: {}
      security:
      - JWT: []
      summary: logout
      tags:
      - Auth
...
  /organizations/{organizationId}/policy-notifications:
    get:
      consumes:
      - application/json
      description: Get PolicyNotifications
      parameters:
      - description: organizationId
        in: path
        name: organizationId
        required: true
        type: string
      - description: pageSize
        in: query
        name: pageSize
        type: string
      - description: pageNumber
        in: query
        name: pageNumber
        type: string
      - description: sortColumn
        in: query
        name: soertColumn
        type: string
      - description: sortOrder
        in: query
        name: sortOrder
        type: string
      - collectionFormat: csv
        description: filters
        in: query
        items:
          type: string
        name: filters
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetPolicyNotificationsResponse'
      security:
      - JWT: []
      summary: Get PolicyNotifications
      tags:
      - PolicyNotifications
  /organizations/{organizationId}/policy-notifications/{policyNotificationId}:
    get:
      consumes:
      - application/json
      description: Get PolicyNotification
      parameters:
      - description: organizationId
        in: path
        name: organizationId
        required: true
        type: string
      - description: policyNotificationId
        in: path
        name: policyNotificationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_openinfradev_tks-api_pkg_domain.GetPolicyNotificationResponse'
      security:
      - JWT: []
      summary: Get PolicyNotification
      tags:
      - PolicyNotifications
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"